Making rule-based system from Pandemic public
- And adding AbstractRuleBasedForwardModel to handle the rule system in a more generic and consistent way
- Documentation for usage included
