+++ src/main/java/core/Game.java
@@ -245,7 +245,7 @@ public class Game {
         ActionController ac = new ActionController();
 
         /* 1. Choose game to play */
-        GameType gameToPlay = ColtExpress;
+        GameType gameToPlay = Pandemic;
 //        List<GameType> games = GameType.Mechanic.Cooperative.getAllGames();
 
         /* 2. Running with visuals? */
