+++ src/main/java/core/rules/rulenodes/ForceAllPlayerReaction.java
@@ -1,11 +1,16 @@
-package games.pandemic.engine.rules;
+package core.rules.rulenodes;
 
 import core.AbstractGameState;
+import core.rules.nodetypes.RuleNode;
 import core.turnorders.ReactiveTurnOrder;
 
 import static core.AbstractGameState.DefaultGamePhase.PlayerReaction;
 
-public class ForceAllEventReaction extends RuleNode {
+/**
+ * Forces all players to play a reaction, in games where a ReactiveTurnOrder is used.
+ * Should be followed by PlayerAction rule nodes, one for each player, in sequence.
+ */
+public class ForceAllPlayerReaction extends RuleNode {
 
     @Override
     protected boolean run(AbstractGameState gs) {
