+++ src/main/java/core/rules/rulenodes/PlayerAction.java
@@ -1,0 +1,25 @@
+package core.rules.rulenodes;
+
+import core.AbstractGameState;
+import core.rules.nodetypes.RuleNode;
+
+/**
+ * Executes an action requested by a player, if the action was set. Interrupts the game loop if no action was provided,
+ * but not if the action fails execution.
+ */
+public class PlayerAction extends RuleNode {
+
+    public PlayerAction() {
+        super(true);
+    }
+
+    @Override
+    protected boolean run(AbstractGameState gs) {
+        if (action != null) {
+            action.execute(gs);
+            return true;
+        }
+        return false;
+    }
+
+}
