+++ src/main/java/games/pandemic/rules/conditions/ActionsPerTurnPlayed.java
@@ -1,6 +1,7 @@
-package games.pandemic.engine.conditions;
+package games.pandemic.rules.conditions;
 
 import core.AbstractGameState;
+import core.rules.nodetypes.ConditionNode;
 import games.pandemic.PandemicTurnOrder;
 
 public class ActionsPerTurnPlayed extends ConditionNode {
