+++ src/main/java/games/pandemic/rules/conditions/HasRPCard.java
@@ -1,36 +1,31 @@
-package games.pandemic.engine.rules;
+package games.pandemic.rules.conditions;
 
 import core.AbstractGameState;
-import games.pandemic.PandemicTurnOrder;
 import core.components.Card;
 import core.components.Deck;
 import core.properties.PropertyString;
+import core.rules.nodetypes.ConditionNode;
 import games.pandemic.PandemicGameState;
 
-import static games.pandemic.PandemicGameState.PandemicGamePhase.RPReaction;
 import static core.CoreConstants.nameHash;
 import static core.CoreConstants.playerHandHash;
 
 @SuppressWarnings("unchecked")
-public class ForceRPReaction extends RuleNode {
+public class HasRPCard extends ConditionNode {
 
     @Override
-    protected boolean run(AbstractGameState gs) {
-        PandemicGameState pgs = (PandemicGameState)gs;
+    protected boolean test(AbstractGameState gs) {
         int nPlayers = gs.getNPlayers();
-
         for (int i = 0; i < nPlayers; i++) {
-            Deck<Card> ph = (Deck<Card>) pgs.getComponent(playerHandHash, i);
+            Deck<Card> ph = (Deck<Card>) ((PandemicGameState)gs).getComponent(playerHandHash, i);
             int nCards = ph.getSize();
             for (int cp = 0; cp < nCards; cp++) {
                 Card card = ph.getComponents().get(cp);
                 if (((PropertyString)card.getProperty(nameHash)).value.equals("Resilient Population")) {
-                    ((PandemicTurnOrder)pgs.getTurnOrder()).addReactivePlayer(i);
-                    pgs.setGamePhase(RPReaction);
-                    return false;
+                    return true;
                 }
             }
         }
-        return true;
+        return false;
     }
 }
