+++ src/main/java/games/pandemic/rules/rules/EpidemicInfect.java
@@ -1,9 +1,10 @@
-package games.pandemic.engine.rules;
+package games.pandemic.rules.rules;
 
 import core.AbstractGameState;
 import core.components.Card;
 import core.components.Counter;
 import core.components.Deck;
+import core.rules.nodetypes.RuleNode;
 import games.pandemic.PandemicGameState;
 import games.pandemic.actions.InfectCity;
