+++ src/main/java/games/pandemic/rules/rules/ForceDiscardReaction.java
@@ -1,8 +1,9 @@
-package games.pandemic.engine.rules;
+package games.pandemic.rules.rules;
 
 import core.AbstractGameState;
 import core.components.Card;
 import core.components.Deck;
+import core.rules.nodetypes.RuleNode;
 import games.pandemic.PandemicGameState;
 import games.pandemic.PandemicTurnOrder;
