+++ src/main/java/games/pandemic/rules/rules/ForceRPReaction.java
@@ -1,30 +1,37 @@
-package games.pandemic.engine.conditions;
+package games.pandemic.rules.rules;
 
 import core.AbstractGameState;
+import core.rules.nodetypes.RuleNode;
+import games.pandemic.PandemicTurnOrder;
 import core.components.Card;
 import core.components.Deck;
 import core.properties.PropertyString;
 import games.pandemic.PandemicGameState;
 
+import static games.pandemic.PandemicGameState.PandemicGamePhase.RPReaction;
 import static core.CoreConstants.nameHash;
 import static core.CoreConstants.playerHandHash;
 
 @SuppressWarnings("unchecked")
-public class HasRPCard extends ConditionNode {
+public class ForceRPReaction extends RuleNode {
 
     @Override
-    protected boolean test(AbstractGameState gs) {
+    protected boolean run(AbstractGameState gs) {
+        PandemicGameState pgs = (PandemicGameState)gs;
         int nPlayers = gs.getNPlayers();
+
         for (int i = 0; i < nPlayers; i++) {
-            Deck<Card> ph = (Deck<Card>) ((PandemicGameState)gs).getComponent(playerHandHash, i);
+            Deck<Card> ph = (Deck<Card>) pgs.getComponent(playerHandHash, i);
             int nCards = ph.getSize();
             for (int cp = 0; cp < nCards; cp++) {
                 Card card = ph.getComponents().get(cp);
                 if (((PropertyString)card.getProperty(nameHash)).value.equals("Resilient Population")) {
-                    return true;
+                    ((PandemicTurnOrder)pgs.getTurnOrder()).addReactivePlayer(i);
+                    pgs.setGamePhase(RPReaction);
+                    return false;
                 }
             }
         }
-        return false;
+        return true;
     }
 }
