+++ src/main/java/games/pandemic/rules/rules/InfectCities.java
@@ -1,10 +1,11 @@
-package games.pandemic.engine.rules;
+package games.pandemic.rules.rules;
 
 import core.AbstractGameState;
 import core.actions.DrawCard;
 import core.components.Card;
 import core.components.Counter;
 import core.components.Deck;
+import core.rules.nodetypes.RuleNode;
 import games.pandemic.PandemicGameState;
 import games.pandemic.actions.InfectCity;
